ffmpeg-qsv: move common code to lib
